+++ app/src/main/java/org/eyeseetea/malariacare/data/remote/sdk/PullDhisSDKDataSource.java
@@ -28,7 +28,8 @@ import com.raizlabs.android.dbflow.sql.language.Delete;
 import com.raizlabs.android.dbflow.sql.language.SQLite;
 
 import org.eyeseetea.malariacare.data.IPullSourceCallback;
-import org.eyeseetea.malariacare.data.database.iomodules.dhis.importer.PullController;
+import org.eyeseetea.malariacare.data.database.iomodules.dhis.importer.PullDataController;
+import org.eyeseetea.malariacare.data.database.iomodules.dhis.importer.PullMetadataController;
 import org.eyeseetea.malariacare.data.database.utils.PreferencesState;
 import org.eyeseetea.malariacare.data.remote.IPullDataSource;
 import org.eyeseetea.malariacare.domain.exception.NetworkException;
@@ -84,7 +85,7 @@ public class PullDhisSDKDataSource implements IPullDataSource {
         } else {
             Set<ProgramType> programTypes = new HashSet<>();
             programTypes.add(ProgramType.WITHOUT_REGISTRATION);
-            if (!PullController.PULL_IS_ACTIVE) {
+            if (!PullMetadataController.PULL_IS_ACTIVE) {
                 return;
             }
             Scheduler pullThread = Schedulers.newThread();
@@ -126,9 +127,9 @@ public class PullDhisSDKDataSource implements IPullDataSource {
             callback.onError(new NetworkException());
         } else {
             try {
-                if (!PullController.PULL_IS_ACTIVE) return;
+                if (!PullMetadataController.PULL_IS_ACTIVE) return;
                 pullEvents(filters, callback);
-                if (!PullController.PULL_IS_ACTIVE) return;
+                if (!PullMetadataController.PULL_IS_ACTIVE) return;
                 callback.onComplete();
             } catch (Exception e) {
                 callback.onError(e);
@@ -144,12 +145,12 @@ public class PullDhisSDKDataSource implements IPullDataSource {
                 D2.me().organisationUnits().list().subscribeOn(listThread)
                         .observeOn(listThread).toBlocking().single();
 
-        if (!PullController.PULL_IS_ACTIVE) return;
+        if (!PullDataController.PULL_IS_ACTIVE) return;
         for (Program program : sdkPrograms) {
             for (OrganisationUnit organisationUnit : sdkOrganisationUnits) {
                 for (Program orgunitProgram : organisationUnit.getPrograms()) {
                     if (orgunitProgram.getUId().equals(program.getUId())) {
-                        if (!PullController.PULL_IS_ACTIVE) return;
+                        if (!PullDataController.PULL_IS_ACTIVE) return;
 
                         EventFilters eventFilters = new EventFilters();
                         eventFilters.setProgramUId(program.getUId());
