+++ app/src/main/java/org/eyeseetea/malariacare/domain/boundary/IPullDataController.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright (c) 2017.
+ *
+ * This file is part of QA App.
+ *
+ *  Health Network QIS App is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  Health Network QIS App is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with Foobar.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package org.eyeseetea.malariacare.domain.boundary;
+
+import org.eyeseetea.malariacare.domain.usecase.pull.PullFilters;
+import org.eyeseetea.malariacare.domain.usecase.pull.PullStep;
+
+public interface IPullDataController {
+    interface Callback {
+        void onComplete();
+
+        void onStep(PullStep step);
+
+        void onError(Throwable throwable);
+
+        void onCancel();
+    }
+
+    void pullData(PullFilters filters, Callback callback);
+
+    void cancel();
+
+    boolean isPullActive();
+}
